Major namespace refactoring and additional value object classes
