+++ src/Authentication/Constants/ChallengeMismatchException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Pinnacle\OpenIdConnect\Authentication\Constants;
+
+use Pinnacle\OpenIdConnect\Support\Exceptions\OpenIdConnectException;
+
+class ChallengeMismatchException extends OpenIdConnectException
+{
+}
