+++ src/Authentication/Models/Scope.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Pinnacle\OpenIdConnect\Authentication\Models;
+
+use Pinnacle\OpenIdConnect\Support\Models\NonEmptyString;
+
+class Scope extends NonEmptyString
+{
+}
