+++ src/Authentication/StatePersister/Exceptions/StatePersisterMissingValueException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Pinnacle\OpenIdConnect\Authentication\StatePersister\Exceptions;
+
+use Pinnacle\OpenIdConnect\Support\Exceptions\OpenIdConnectException;
+
+class StatePersisterMissingValueException extends OpenIdConnectException
+{
+}
