+++ src/Authorization/AuthorizationCodeResponse.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace Pinnacle\OpenIdConnect\Authorization;
+
+use GuzzleHttp\Psr7\Uri;
+use Pinnacle\OpenIdConnect\Authentication\Models\Challenge;
+use Pinnacle\OpenIdConnect\Provider\Contracts\ProviderConfigurationInterface;
+use Pinnacle\OpenIdConnect\Provider\Models\ProviderConfiguration;
+
+class AuthorizationCodeResponse
+{
+    public function __construct(
+        private string                $authorizationCode,
+        private ProviderConfiguration $provider,
+        private Uri                   $redirectUri,
+        private Challenge             $challenge
+    ) {
+    }
+
+    public function getProvider(): ProviderConfigurationInterface
+    {
+        return $this->provider;
+    }
+
+    public function getAuthorizationCode(): string
+    {
+        return $this->authorizationCode;
+    }
+
+    public function getRedirectUri(): Uri
+    {
+        return $this->redirectUri;
+    }
+
+    public function getChallenge(): Challenge
+    {
+        return $this->challenge;
+    }
+}
