--- src/Exceptions/AccessTokenNotFoundException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Pinnacle\OpenIdConnect\Exceptions;
-
-class AccessTokenNotFoundException extends OpenIdConnectException
-{
-}
