--- src/Exceptions/ChallengeMismatchException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Pinnacle\OpenIdConnect\Exceptions;
-
-class ChallengeMismatchException extends OpenIdConnectException
-{
-}
