--- src/Exceptions/InsecureUriException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Pinnacle\OpenIdConnect\Exceptions;
-
-class InsecureUriException extends OpenIdConnectException
-{
-}
