--- src/Exceptions/OpenIdConnectException.php
@@ -1,14 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Pinnacle\OpenIdConnect\Exceptions;
-
-use DomainException;
-
-/**
- * An exception thrown when OAuth authentication fails.
- */
-class OpenIdConnectException extends DomainException
-{
-}
