--- src/Provider/ClientId.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Pinnacle\OpenIdConnect\Provider;
-
-use Pinnacle\OpenIdConnect\Support\NonEmptyString;
-
-class ClientId extends NonEmptyString
-{
-}
