--- src/Provider/Identifier.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Pinnacle\OpenIdConnect\Provider;
-
-use Assert\Assert;
-
-class Identifier
-{
-    public function __construct(private string|int $identifierValue)
-    {
-        if (is_string($this->identifierValue)) {
-            Assert::that($this->identifierValue)->notBlank();
-        }
-    }
-
-    public function getValue(): string|int
-    {
-        return $this->identifierValue;
-    }
-}
