+++ src/Provider/Models/ClientId.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Pinnacle\OpenIdConnect\Provider\Models;
+
+use Pinnacle\OpenIdConnect\Support\Models\NonEmptyString;
+
+class ClientId extends NonEmptyString
+{
+}
