+++ src/Provider/Models/ClientSecret.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Pinnacle\OpenIdConnect\Provider\Models;
+
+use Pinnacle\OpenIdConnect\Support\Models\NonEmptyString;
+
+class ClientSecret extends NonEmptyString
+{
+}
