+++ src/Requestors/Exceptions/AccessTokenNotFoundException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Pinnacle\OpenIdConnect\Requestors\Exceptions;
+
+use Pinnacle\OpenIdConnect\Support\Exceptions\OpenIdConnectException;
+
+class AccessTokenNotFoundException extends OpenIdConnectException
+{
+}
