--- src/State/Exceptions/StatePersisterMissingValueException.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Pinnacle\OpenIdConnect\State\Exceptions;
-
-use Pinnacle\OpenIdConnect\Exceptions\OpenIdConnectException;
-
-class StatePersisterMissingValueException extends OpenIdConnectException
-{
-}
