+++ src/Support/Exceptions/InsecureUriException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Pinnacle\OpenIdConnect\Support\Exceptions;
+
+class InsecureUriException extends OpenIdConnectException
+{
+}
