+++ src/Support/Exceptions/OpenIdConnectException.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Pinnacle\OpenIdConnect\Support\Exceptions;
+
+use DomainException;
+
+/**
+ * An exception thrown when OAuth authentication fails.
+ */
+class OpenIdConnectException extends DomainException
+{
+}
