+++ src/Support/Models/NonEmptyString.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Pinnacle\OpenIdConnect\Support\Models;
+
+use Assert\Assert;
+
+class NonEmptyString
+{
+    public function __construct(private string $value)
+    {
+        Assert::that($this->value)->notBlank();
+    }
+
+    public function getValue(): string
+    {
+        return $this->value;
+    }
+}
